Add --read-requirements option to check the requirements.txt
This can be used to ensure the requirements.txt doesn't allow
insecure versions.

This only works for lines which actually specify a version.

This is only tested with 'pkg>=1.2.3' like lines.

With small modifications this could be made to parse `pip freeze`
output on stdin.

Related Issue: https://github.com/pyupio/safety/issues/2