Added example to README
Add a small snippet that scans an existing docker image. This ensures that the check
is done on the final installation - which may be deployed in production - as opposed
to the requirements.txt, which often (i.e. in the wild) does not contain all dependencies 
or the complete state of the installed dependencies (e.g. wheel, setuptools).